requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
cryptography==44.0.3
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.4.2
debugpy==1.8.19
decorator==5.2.1
dill==0.4.0
distro==1.9.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
executing==2.2.1
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jiwer==4.0.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.7.0
jupyter_core==5.9.1
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.4
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
llvmlite==0.46.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai==2.14.0
openai-whisper==20250625
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
phonenumbers==9.0.21
pika==1.3.2
pillow==12.0.0
platformdirs==4.5.1
preshed==3.0.12
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
psycopg2==2.9.11
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
scikit-learn==1.8.0
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tldextract==5.3.0
torch==2.9.1+cu130
torchcodec==0.9.1
torchvision==0.24.1+cu130
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.3
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
